♻️ TableEditor: extracted SearchBox
