+++ lib/TableEditor/index.jsx
@@ -16,15 +16,14 @@ import Typography from '@mui/material/Typography'
 import Paper from '@mui/material/Paper'
 import Checkbox from '@mui/material/Checkbox'
 import IconButton from '@mui/material/IconButton'
-import SearchIcon from '@mui/icons-material/Search'
 import Tooltip from '@mui/material/Tooltip'
 import Collapse from '@mui/material/Collapse'
 import Button from '@mui/material/Button'
 import { visuallyHidden } from '@mui/utils'
-import InputBase from '@mui/material/InputBase'
-import { styled, alpha } from '@mui/material/styles'
+import { alpha } from '@mui/material/styles'
 import { useTranslation } from 'react-i18next'
 import Loading from './Loading'
+import SearchBox from './SearchBox'
 import i18n from '../i18n'
 
 /* eslint-enable */
@@ -70,48 +69,6 @@ const ActionsType = PropTypes.arrayOf(
 ActionButtons.propTypes = {
   actions: ActionsType,
 }
-
-const Search = styled('div')(({ theme }) => ({
-  position: 'relative',
-  borderRadius: theme.shape.borderRadius,
-  backgroundColor: alpha(theme.palette.common.white, 0.15),
-  '&:hover': {
-    backgroundColor: alpha(theme.palette.common.white, 0.25),
-  },
-  marginLeft: 0,
-  width: '100%',
-  [theme.breakpoints.up('sm')]: {
-    marginLeft: theme.spacing(1),
-    width: 'auto',
-  },
-}))
-
-const SearchIconWrapper = styled('div')(({ theme }) => ({
-  padding: theme.spacing(0, 2),
-  height: '100%',
-  position: 'absolute',
-  pointerEvents: 'none',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-}))
-
-const StyledInputBase = styled(InputBase)(({ theme }) => ({
-  color: 'inherit',
-  '& .MuiInputBase-input': {
-    padding: theme.spacing(1, 1, 1, 0),
-    // vertical padding + font size from searchIcon
-    paddingLeft: `calc(1em + ${theme.spacing(4)})`,
-    transition: theme.transitions.create('width'),
-    width: '100%',
-    [theme.breakpoints.up('sm')]: {
-      width: '12ch',
-      '&:focus': {
-        width: '20ch',
-      },
-    },
-  },
-}))
 
 function descendingComparator(a, b, orderBy) {
   function define(v) {
@@ -229,7 +186,7 @@ function EnhancedTableToolbar({
   title,
   selected,
   numSelected,
-  onSearchEdited,
+  setSearch,
   search,
   actions,
   selectionActions,
@@ -262,17 +219,7 @@ function EnhancedTableToolbar({
       )}
 
       <Box variant="h6" id="Filter" component="div">
-        <Search>
-          <SearchIconWrapper>
-            <SearchIcon />
-          </SearchIconWrapper>
-          <StyledInputBase
-            placeholder={t('TABLE_EDITOR.PLACEHOLDER_SEARCH')}
-            inputProps={{ 'aria-label': t('TABLE_EDITOR.LABEL_SEARCH') }}
-            onChange={onSearchEdited}
-            value={search}
-          />
-        </Search>
+        <SearchBox {...{ setSearch, search }} />
       </Box>
       <Box sx={{ flex: '1 1 100%' }}></Box>
       <ActionButtons
@@ -288,7 +235,7 @@ EnhancedTableToolbar.propTypes = {
   numSelected: PropTypes.number.isRequired,
   search: PropTypes.string,
   selected: PropTypes.arrayOf(PropTypes.string),
-  onSearchEdited: PropTypes.func,
+  setSearch: PropTypes.func,
   actions: ActionsType,
   selectionActions: ActionsType,
 }
@@ -413,9 +360,7 @@ function TableEditor(props) {
           numSelected={selected.length}
           selected={selected}
           search={search}
-          onSearchEdited={(ev) => {
-            setSearch(ev.target.value)
-          }}
+          setSearch={setSearch}
           actions={actions}
           selectionActions={selectionActions}
         />
